Changed value in Checkstyle configuration
Due to the diversity of situations in which this warning was detected, it was considered a false-positive.
An example is the situations such as long method declarations could be solved by renaming method parameters but that would maybe incur in confusing or unclear variable names.

Value for max number of characters per line for detection was changed to 150.